Minor changes to where tree was stored
